+++ src/com/example/java/PokemonTester.java
@@ -2,7 +2,6 @@ package com.example.java;
 
 import javax.swing.*;
 
-import static com.example.java.Pokemon.battle;
 
 import static com.example.java.Starter.*;
 
@@ -10,12 +9,10 @@ import static com.example.java.Starter.*;
 
 public class PokemonTester {
 
-    public static WildPokemon wildPokemon;
-
 
     public static void main(String[] args) {
         Starter.selectStarter();
-        Starter.getNewStarter(newStarter);
+        Starter.createNewStarter(newStarter);
         Starter starter = new Starter(newStarter, maxStarterHealth, starterAttack, starterSpeed);
 
         WildPokemon wildPokemon = new WildPokemon();
@@ -27,10 +24,10 @@ public class PokemonTester {
                 "Pokemon Adventure", JOptionPane.QUESTION_MESSAGE, null, selectionValues, initialSelection);
         if (selection == "Zone 1 (Easy)") {
             JOptionPane.showMessageDialog(null, "You have chosen Zone 1!");
-            battle(starter, wildPokemon.getRandomPokemonZone1());
+            new Battle(starter, wildPokemon.getRandomPokemonZone1());
         } else if (selection == "Zone 2 (Medium)") {
             JOptionPane.showMessageDialog(null, "You have chosen Zone 2!");
-            battle(starter, wildPokemon.getRandomPokemonZone2());
+            new Battle(starter, wildPokemon.getRandomPokemonZone2());
         }
     }
 }
