+++ src/com/example/java/WildPokemon.java
@@ -7,45 +7,34 @@ import java.util.Random;
 class WildPokemon {
 
 
-
-    private int randomHealthZone1(int min, int max) {
+    private int randomHealth(int min, int max) {
         int range = (max - min) + 1;
         return (int)(Math.random() * range) + min;
     }
-    private int randomAttackZone1(int min, int max) {
+    private int randomAttack(int min, int max) {
         int range = (max - min) + 1;
         return (int)(Math.random() * range) + min;
     }
-    private int randomSpeedZone1(int min, int max) {
-        int range = (max - min) + 1;
-        return (int)(Math.random() * range) + min;
-    }
-
-    private int randomHealthZone2(int min, int max) {
-        int range = (max - min) + 1;
-        return (int)(Math.random() * range) + min;
-    }
-    private int randomAttackZone2(int min, int max) {
-        int range = (max - min) + 1;
-        return (int)(Math.random() * range) + min;
-    }
-    private int randomSpeedZone2(int min, int max) {
+    private int randomSpeed(int min, int max) {
         int range = (max - min) + 1;
         return (int)(Math.random() * range) + min;
     }
 
-    private Random rand = new Random();
-    private List<Pokemon> pokemonZone1;
-    private List<Pokemon> pokemonZone2;
+
+    public Random rand = new Random();
+    public List<Pokemon> pokemonZone1;
+    public List<Pokemon> pokemonZone2;
 
     WildPokemon(){
         pokemonZone1 = new ArrayList();
         pokemonZone2 = new ArrayList();
 
-        Pokemon rattata = new Pokemon("Rattata",randomHealthZone1(15,20),randomAttackZone1(2,5),randomSpeedZone1(3,6), true);
+        Pokemon rattata = new Pokemon("Rattata",randomHealth(15,20),randomAttack(2,5),randomSpeed(3,6), true);
         pokemonZone1.add(rattata);
-        Pokemon pidgey = new Pokemon("Pidgey",randomHealthZone1(10,17),randomAttackZone1(3,4),randomSpeedZone1(2,5), true);
+        Pokemon pidgey = new Pokemon("Pidgey",randomHealth(12,17),randomAttack(3,4),randomSpeed(2,5), true);
         pokemonZone1.add(pidgey);
+        Pokemon caterpie = new Pokemon("Caterpie",randomHealth(20,23),randomAttack(3,4),randomSpeed(2,5), true);
+        pokemonZone2.add(caterpie);
 
 }
 
@@ -58,6 +47,7 @@ class WildPokemon {
         int n = rand.nextInt(pokemonZone2.size());
         return pokemonZone2.get(n);
     }
+
 /*
     public static void resetStatsZone1() {
         Pokemon.health = randomHealthZone1(15,20);
